+++ resources/views/components/ui/select-content.blade.php
@@ -6,7 +6,7 @@
 
 @php
     $placement = $side.($align === 'center' ? '' : '-'.$align);
-    $anchorAttr = 'x-anchor.'.$placement.'.offset.'.$sideOffset.'="$refs.trigger"';
+    $anchorAttr = 'x-blat-anchor.'.$placement.'.offset.'.$sideOffset.'="$refs.trigger"';
 @endphp
 
 <template x-teleport="body" wire:ignore>